server/requirements.txt
celery[redis]
fastapi[standard]
firebase-admin
requests
python-dotenv
uvicorn[standard]
serpapi
google-search-results
